insertImage.tsの関数をAPIのエンドポイントごとに作成した
